requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
imageio==2.31.3
lazy_loader==0.3
llvmlite==0.40.1
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.3
PyMatting==1.1.8
PyWavelets==1.4.1
rembg==2.0.50
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.2
sympy==1.12
tifffile==2023.8.30
tqdm==4.66.1
urllib3==2.0.4
